test(api): split route coverage by module
