+++ packages/api/src/routes/admin.test.ts
@@ -1,0 +1,1 @@
+import "../test-workspace-root.js";
