+++ packages/api/src/routes/goals.test.ts
@@ -1,0 +1,1 @@
+import "../test-workspace-root.js";
